Move models to FridgeModels.kt
